pyproject.toml
[tool.poetry]
name = "image_color_areas"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.3"
Pillow = "^9.1.0"
Flask = "^2.1.2"
Flask-WTF = "^1.0.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pyreadline3 = "^3.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
